requirements.txt
aniso8601
click
Flask
Flask-RESTful
itsdangerous
Jinja2
MarkupSafe
python-dateutil
pytz
uwsgi
PyYaml